Moving to event ingestion service and client libs
- Removed the publisher module
+ Moved to java 11
+ Moved to dropwizard 2x
+ Moved to ingestion client to ingest events
+ Initialized docker mariadb instances properly using sql we have in the project
