+++ statesman-engine/src/main/java/io/appform/statesman/engine/action/BaseAction.java
@@ -7,6 +7,8 @@ import com.github.rholder.retry.Retryer;
 import com.github.rholder.retry.RetryerBuilder;
 import com.github.rholder.retry.StopStrategies;
 import com.github.rholder.retry.WaitStrategies;
+import io.appform.eventingester.client.EventPublisher;
+import io.appform.eventingester.models.Event;
 import io.appform.statesman.engine.Constants;
 import io.appform.statesman.engine.events.ActionExecutedEvent;
 import io.appform.statesman.engine.events.EngineEventType;
@@ -14,8 +16,6 @@ import io.appform.statesman.model.Action;
 import io.appform.statesman.model.Workflow;
 import io.appform.statesman.model.action.template.ActionTemplate;
 import io.appform.statesman.model.exception.StatesmanError;
-import io.appform.statesman.publisher.EventPublisher;
-import io.appform.statesman.publisher.model.Event;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.Collections;
@@ -33,7 +33,7 @@ public abstract class BaseAction<T extends ActionTemplate> implements Action<T> 
     protected final ObjectMapper mapper;
     private final Retryer<JsonNode> retryer;
 
-    public BaseAction(EventPublisher publisher, ObjectMapper mapper) {
+    protected BaseAction(EventPublisher publisher, ObjectMapper mapper) {
         this.publisher = publisher;
         this.mapper = mapper;
         retryer = RetryerBuilder.<JsonNode>newBuilder()
@@ -65,7 +65,7 @@ public abstract class BaseAction<T extends ActionTemplate> implements Action<T> 
     private void publish(final List<Event> eventList) {
         try {
             if (null != eventList && !eventList.isEmpty()) {
-                this.publisher.publish(Constants.FOXTROT_REPORTING_TOPIC, eventList);
+                this.publisher.publish(eventList);
             }
         } catch (final Exception e) {
             log.error("Unable to send event", e);
