+++ statesman-engine/src/main/java/io/appform/statesman/engine/action/impl/EvaluatedAction.java
@@ -7,6 +7,7 @@ import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.LoadingCache;
 import com.google.common.base.Preconditions;
 import com.google.inject.name.Named;
+import io.appform.eventingester.client.EventPublisher;
 import io.appform.hope.core.Evaluatable;
 import io.appform.hope.core.exceptions.errorstrategy.InjectValueErrorHandlingStrategy;
 import io.appform.hope.lang.HopeLangEngine;
@@ -18,8 +19,6 @@ import io.appform.statesman.model.Workflow;
 import io.appform.statesman.model.action.ActionType;
 import io.appform.statesman.model.action.template.EvaluatedActionTemplate;
 import io.appform.statesman.model.action.template.RuleBasedTemplate;
-import io.appform.statesman.model.exception.StatesmanError;
-import io.appform.statesman.publisher.EventPublisher;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
