+++ statesman-engine/src/main/java/io/appform/statesman/engine/action/impl/HttpAction.java
@@ -5,16 +5,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.NullNode;
 import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
+import io.appform.eventingester.client.EventPublisher;
 import io.appform.statesman.engine.action.BaseAction;
 import io.appform.statesman.engine.handlebars.HandleBarsService;
+import io.appform.statesman.engine.http.HttpClient;
+import io.appform.statesman.engine.http.HttpUtil;
 import io.appform.statesman.model.ActionImplementation;
 import io.appform.statesman.model.Workflow;
 import io.appform.statesman.model.action.ActionType;
 import io.appform.statesman.model.action.template.HttpActionTemplate;
 import io.appform.statesman.model.exception.StatesmanError;
-import io.appform.statesman.publisher.EventPublisher;
-import io.appform.statesman.publisher.http.HttpClient;
-import io.appform.statesman.publisher.http.HttpUtil;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 import okhttp3.Response;
