+++ statesman-engine/src/main/java/io/appform/statesman/engine/http/HttpUtil.java
@@ -1,4 +1,4 @@
-package io.appform.statesman.publisher.http;
+package io.appform.statesman.engine.http;
 
 import com.codahale.metrics.MetricRegistry;
 import com.raskasa.metrics.okhttp.InstrumentedOkHttpClients;
