+++ statesman-model/pom.xml
@@ -18,9 +18,9 @@
             <version>${dropwizard.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-validator</artifactId>
-            <version>${hibernate-validator.version}</version>
+            <groupId>io.dropwizard</groupId>
+            <artifactId>dropwizard-validation</artifactId>
+            <version>${dropwizard.version}</version>
         </dependency>
     </dependencies>
 </project>
