+++ statesman-server/pom.xml
@@ -12,10 +12,9 @@
     <artifactId>statesman-server</artifactId>
 
     <properties>
-        <guice.version>4.2.0</guice.version>
-        <dropwizard.swagger.version>1.3.7-1</dropwizard.swagger.version>
-        <dropwizard.riemann.version>1.3.12-1</dropwizard.riemann.version>
-        <dropwizard.sharding.version>1.3.13-6</dropwizard.sharding.version>
+        <dw-guicy.version>5.3.0</dw-guicy.version>
+        <dropwizard.swagger.version>2.0.12-1</dropwizard.swagger.version>
+        <dropwizard.sharding.version>2.0.23-2</dropwizard.sharding.version>
         <mariadb.driver.version>2.2.0</mariadb.driver.version>
     </properties>
 
@@ -33,7 +32,7 @@
         <dependency>
             <groupId>ru.vyarus</groupId>
             <artifactId>dropwizard-guicey</artifactId>
-            <version>${guice.version}</version>
+            <version>${dw-guicy.version}</version>
         </dependency>
         <dependency>
             <groupId>io.dropwizard</groupId>
@@ -108,14 +107,25 @@
         </dependency>
 
         <dependency>
-            <groupId>io.appform</groupId>
-            <artifactId>statesman-publisher</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>io.appform.functionmetrics</groupId>
             <artifactId>function-metrics</artifactId>
             <version>${function-metrics.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.aspectj</groupId>
+                    <artifactId>*</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjrt</artifactId>
+            <version>${aspectj.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+            <version>${aspectj.version}</version>
         </dependency>
         <dependency>
             <groupId>com.github.tomakehurst</groupId>
@@ -177,24 +187,24 @@
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>aspectj-maven-plugin</artifactId>
-                <version>1.11</version>
+                <version>1.14.0</version>
                 <dependencies>
                     <dependency>
                         <groupId>org.aspectj</groupId>
                         <artifactId>aspectjrt</artifactId>
-                        <version>1.8.13</version>
+                        <version>1.9.7</version>
                     </dependency>
                     <dependency>
                         <groupId>org.aspectj</groupId>
                         <artifactId>aspectjtools</artifactId>
-                        <version>1.8.13</version>
+                        <version>1.9.7</version>
                     </dependency>
                 </dependencies>
 
                 <configuration>
-                    <complianceLevel>1.8</complianceLevel>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <complianceLevel>11</complianceLevel>
+                    <source>11</source>
+                    <target>11</target>
                     <showWeaveInfo>true</showWeaveInfo>
                     <forceAjcCompile>true</forceAjcCompile>
                     <sources/>
