+++ statesman-server/src/main/java/io/appform/statesman/server/idextractor/FqlIdExtractor.java
@@ -7,8 +7,8 @@ import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
 import io.appform.statesman.engine.handlebars.HandleBarsService;
 import io.appform.statesman.model.FoxtrotClientConfig;
-import io.appform.statesman.publisher.http.HttpClient;
-import io.appform.statesman.publisher.http.HttpUtil;
+import io.appform.statesman.engine.http.HttpClient;
+import io.appform.statesman.engine.http.HttpUtil;
 import io.appform.statesman.server.callbacktransformation.TransformationTemplate;
 import javax.ws.rs.NotFoundException;
 import lombok.SneakyThrows;
