+++ statesman-server/src/main/java/io/appform/statesman/server/module/StatesmanModule.java
@@ -5,6 +5,8 @@ import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import com.google.inject.name.Named;
 import com.google.inject.name.Names;
+import io.appform.eventingester.client.EventPublisher;
+import io.appform.eventingester.client.EventPublishers;
 import io.appform.statesman.engine.ActionTemplateStore;
 import io.appform.statesman.engine.ProviderSelector;
 import io.appform.statesman.engine.TransitionStore;
@@ -13,19 +15,14 @@ import io.appform.statesman.engine.action.ActionExecutor;
 import io.appform.statesman.engine.action.ActionExecutorImpl;
 import io.appform.statesman.engine.action.ActionRegistry;
 import io.appform.statesman.engine.action.MapBasedActionRegistry;
+import io.appform.statesman.engine.http.HttpClient;
+import io.appform.statesman.engine.http.HttpUtil;
 import io.appform.statesman.engine.observer.ObservableEventBus;
 import io.appform.statesman.engine.observer.ObservableEventBusSubscriber;
 import io.appform.statesman.engine.observer.ObservableGuavaEventBus;
 import io.appform.statesman.engine.observer.observers.FoxtrotEventSender;
 import io.appform.statesman.model.FoxtrotClientConfig;
 import io.appform.statesman.model.HttpClientConfiguration;
-import io.appform.statesman.model.exception.StatesmanError;
-import io.appform.statesman.publisher.EventPublisher;
-import io.appform.statesman.publisher.http.HttpClient;
-import io.appform.statesman.publisher.http.HttpUtil;
-import io.appform.statesman.publisher.impl.QueueEventPublisher;
-import io.appform.statesman.publisher.impl.SyncEventPublisher;
-import io.appform.statesman.publisher.model.PublisherType;
 import io.appform.statesman.server.AppConfig;
 import io.appform.statesman.server.dao.action.ActionTemplateStoreCommand;
 import io.appform.statesman.server.dao.callback.CallbackTemplateProvider;
@@ -38,8 +35,6 @@ import io.appform.statesman.server.idextractor.CompoundIdExtractor;
 import io.appform.statesman.server.idextractor.IdExtractor;
 import io.appform.statesman.server.provider.ProviderSelectorImpl;
 import io.dropwizard.setup.Environment;
-
-import java.io.IOException;
 
 public class StatesmanModule extends AbstractModule {
 
@@ -66,30 +61,7 @@ public class StatesmanModule extends AbstractModule {
     public EventPublisher provideEventPublisher(
             AppConfig appConfig,
             Environment environment) {
-
-        return appConfig.getEventPublisherConfig().getType().visit(
-                new PublisherType.PublisherTypeVisitor<EventPublisher>() {
-                    @Override
-                    public EventPublisher visitSync() {
-                        return new SyncEventPublisher(
-                                environment.getObjectMapper(),
-                                appConfig.getEventPublisherConfig(),
-                                environment.metrics()
-                        );
-                    }
-
-                    @Override
-                    public EventPublisher visitQueued() {
-                        try {
-                            return new QueueEventPublisher(
-                                    environment.getObjectMapper(),
-                                    appConfig.getEventPublisherConfig(),
-                                    environment.metrics());
-                        } catch (IOException e) {
-                            throw StatesmanError.propagate(e);
-                        }
-                    }
-                });
+        return EventPublishers.create(appConfig.getEventPublisherConfig(), environment.getObjectMapper());
     }
 
     @Singleton
