+++ statesman-server/src/main/java/io/appform/statesman/server/resources/EventResource.java
@@ -4,8 +4,8 @@ import com.codahale.metrics.annotation.Timed;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.google.inject.name.Named;
-import io.appform.statesman.publisher.EventPublisher;
-import io.appform.statesman.publisher.model.Event;
+import io.appform.eventingester.client.EventPublisher;
+import io.appform.eventingester.models.Event;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
