+++ statesman-server/src/test/java/io/appform/statesman/server/idextractor/CompoundIdExtractorTest.java
@@ -8,8 +8,8 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import io.appform.statesman.engine.handlebars.HandleBarsService;
 import io.appform.statesman.model.FoxtrotClientConfig;
 import io.appform.statesman.model.HttpClientConfiguration;
-import io.appform.statesman.publisher.http.HttpClient;
-import io.appform.statesman.publisher.http.HttpUtil;
+import io.appform.statesman.engine.http.HttpClient;
+import io.appform.statesman.engine.http.HttpUtil;
 import io.appform.statesman.server.callbacktransformation.TranslationTemplateType;
 import io.appform.statesman.server.callbacktransformation.impl.OneShotTransformationTemplate;
 import javax.ws.rs.NotFoundException;
