+++ statesman-server/startup.sh
@@ -18,6 +18,13 @@ else
     fi
 fi
 
+if [ -z "${PRESLEEP}" ]; then
+  :
+else
+  echo "Sleeping for ${PRESLEEP} seconds..."
+  sleep ${PRESLEEP}
+fi
+
 EXEC_CMD="java -Ddb.shards=${SHARDS-2} -Dfile.encoding=utf-8 -XX:+${GC_ALGO-UseG1GC} -Xms${JAVA_PROCESS_MIN_HEAP-1g} -Xmx${JAVA_PROCESS_MAX_HEAP-1g} ${JAVA_OPTS} -jar server.jar server ${STATESMAN_CONFIG_FILE}"
 
 echo "Number of database shards: ${SHARDS}"
